Module optics::prelude

Prelude, aimed for blanket import.

Lenses for tuples are only defined for each tuple with arity n <= 4. Normally, tuples with more fields are discouraged for practical use. That said, one can always declare his own ones with the macro declare_lens_from_field.

Structs

_0

Lens for field 0 in tuples.

_1

Lens for field 1 in tuples.

_2

Lens for field 2 in tuples.

_3

Lens for field 3 in tuples.

_Err

Prism for Result::Err.

_Ok

Prism for Result::Ok.

_Some

Prism for Option::Some.