Trait optics::traits::Traversal

pub trait Traversal<T>: Setter<T> where
    Self::View: Sized, {
    fn traverse(&self, s: T, f: &mut dyn FnMut(Self::View));

    fn fold<C>(&self, s: T, init: C, f: impl FnMut(&mut C, Self::View)) -> C { ... }
    fn flatten(&self, s: T) -> Vec<Self::View> { ... }
}

Traversal (and also Fold).

Required Methods

fn traverse(&self, s: T, f: &mut dyn FnMut(Self::View))

Evaluate the action from left to right on each element targeted by a Traversal.

Provided Methods

fn fold<C>(&self, s: T, init: C, f: impl FnMut(&mut C, Self::View)) -> C

Fold every element targeted by this Traversal into a single result.

fn flatten(&self, s: T) -> Vec<Self::View>

Flatten the elements targeted by this Traversal into a Vec.