Crate optics

optics: yet another Haskell optics in Rust.

Modules

concrete

Concrete implementations for optics.

prelude

Prelude, aimed for blanket import.

traits

Traits for polymorphic lens hierarchy.

Macros

declare_affine_traversal

Declare an AffineTraversal from an accessor expression.

declare_lens

Declare a Lens from an accessor expression.

declare_lens_from_field

Declare a Lens from a field name for a struct.

declare_prism_from_variant

Declare a Prism from a variant name for an enum.

impl_affine_traversal

Similar to declare_affine_traversal, but does not define the lens type for you.

impl_lens

Similar to declare_lens, but does not define the lens type for you. Normally mark_infallible should be used together with this macro.

impl_up_from

Implement the lens hierarchy from some specific level.

mark_fallible

Mark an optics as fallible by implementing OpticsFallible.

mark_infallible

Mark an optics as infallible by implementing OpticsFallible.

mark_known_source

Mark an optics to have a single known source type by implementing OpticsKnownSource.

optics

Easy composition for optics. See also Compose.