Module optics::concrete

Concrete implementations for optics.

Structs

Compose

Composed optics of K and L; K is applied first.

Identity

Identity optics.

MapFallible

Optics wrapper for mapping the Success and Error value.

SuccessCompose

Success type for Composed optics.

_Identity

Identity optics, explicit about source and view types.

Enums

ErrorCompose

Error type for Composed optics.

Type Definitions

MapError

Optics wrapper for mapping the Error value.

MapFallibleTo

Optics wrapper for mapping the Success and Error value to some specific type.

MapSuccess

Optics wrapper for mapping the Success value.